refactor: move ProcessingFormDialog to features/processing
